add dev endpoint and refactor test
