+++ tox.ini
@@ -73,6 +73,8 @@ commands =
 [testenv:dev]
 description = run FastAPI service for development
 passenv = EVOLVER_LOAD_FROM_CONFIG_ON_STARTUP
+setenv =
+    EVOLVER_DEV_MODE = 1
 extras = dev
 commands =
     python -m evolver.app.main
